LayoutTests/fast/dom/Window/dom-access-from-closure-window-expected.txt

document.URL: file:///Users/weinig/Code/webkit/OpenSource/LayoutTests/fast/dom/Window/resources/notify-opener-done.html
window.document.URL: file:///Users/weinig/Code/webkit/OpenSource/LayoutTests/fast/dom/Window/resources/notify-opener-done.html
name: child
window.name: child
